refactor(CartPage): simplify cart page by integrating CartForm component and removing unused state and form elements
